# dashboard/backend/output_backend/openapi_specs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Statec Census API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/datasets/{dataset_code}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_datasets__dataset_code__get",
        "parameters": [
          {
            "name": "dataset_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataTableDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_code}/dimensions/{dimension_code}": {
      "get": {
        "tags": [
          "dimensions"
        ],
        "summary": "Get Dimension Detail",
        "operationId": "get_dimension_detail_datasets__dataset_code__dimensions__dimension_code__get",
        "parameters": [
          {
            "name": "dataset_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Code"
            }
          },
          {
            "name": "dimension_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dimension Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimensionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_code}/observations": {
      "get": {
        "tags": [
          "observations"
        ],
        "summary": "List Observations",
        "operationId": "list_observations_datasets__dataset_code__observations_get",
        "parameters": [
          {
            "name": "dataset_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "exclusiveMinimum": 0,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObservationPoint"
                  },
                  "title": "Response List Observations Datasets  Dataset Code  Observations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_code}/aggregates": {
      "get": {
        "tags": [
          "observations"
        ],
        "summary": "Aggregate Dataset",
        "operationId": "aggregate_dataset_datasets__dataset_code__aggregates_get",
        "parameters": [
          {
            "name": "dataset_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Code"
            }
          },
          {
            "name": "dimension",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Dimension"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(?i)(asc|desc)$",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_code}/insights/ageing": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Ageing Insights",
        "operationId": "ageing_insights_datasets__dataset_code__insights_ageing_get",
        "parameters": [
          {
            "name": "dataset_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgeingInsights"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AggregateResponse": {
        "properties": {
          "dataset_code": {
            "type": "string",
            "title": "Dataset Code"
          },
          "dimension_code": {
            "type": "string",
            "title": "Dimension Code"
          },
          "filters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Filters"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AggregateItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "dataset_code",
          "dimension_code",
          "filters",
          "results"
        ],
        "title": "AggregateResponse"
      },
...
      "DataTableDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "dimension_count": {
            "type": "integer",
            "title": "Dimension Count"
          },
          "observation_count": {
            "type": "integer",
            "title": "Observation Count"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/DimensionSummary"
            },
            "type": "array",
            "title": "Dimensions"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "dimension_count",
          "observation_count",
          "dimensions"
        ],
        "title": "DataTableDetail"
      },
...
      "DimensionDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "codelist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codelist Id"
          },
          "category_count": {
            "type": "integer",
            "title": "Category Count"
          },
          "applicable_category_count": {
            "type": "integer",
            "title": "Applicable Category Count"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryRead"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "label",
          "position",
          "category_count",
          "applicable_category_count",
          "categories"
        ],
        "title": "DimensionDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ObservationPoint": {
        "properties": {
          "observation_id": {
            "type": "integer",
            "title": "Observation Id"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "time_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Period"
          },
          "dimensions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Dimensions"
          }
        },
        "type": "object",
        "required": [
          "observation_id",
          "value",
          "dimensions"
        ],
        "title": "ObservationPoint"
      },
...
    }
  }
}